Cambiado filtro FIR a IIR, problemas de memoria

+++ StringHandler.c
@@ -1,14 +1,14 @@
 
 #include <stdlib.h>
 #include <string.h>
-
+#include <xc.h>
 #include "GCODE.h"
 
-int getTokens(char** token,char* str){
+uint8_t getTokens(char** token,char* str){
    char s[2]=" ";
    //strcpy(str,buffer);
    token[0] = strtok(str, s);
-   int i=0;
+   uint8_t i=0;
    /* walk through other tokens */
    while( token[i] != NULL) {
       //if(token[i][strlen(token[i])]=='\n') token[i][strlen(token[i])]=0;
@@ -18,9 +18,9 @@ int getTokens(char** token,char* str){
    return i;
 }
 
-void getComands(Comando_T* comandos, char** tokens, int size){
+void getComands(Comando_T* comandos, char** tokens, uint8_t size){
    char buff[10];
-   for(int i=0;i<size;i++){
+   for(uint8_t i=0;i<size;i++){
       comandos[i].code=tokens[i][0];
       strcpy(buff, &tokens[i][1]);
 	comandos[i].number=strtof(buff,NULL);
